--- Tests/LibGC/module.modulemap
@@ -1,6 +1,0 @@
-module GCTesting {
-    header "TestHeap.h"
-    header "TestInterop.h"
-    requires cplusplus
-    export *
-}
